requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
anthropic
gitpython
pytest
